交換の発火の型を実測に合わせる
実測は Fire(14, 0u, 1u) で、index と数量は UInt だった。
int のまま渡していたため AtkValueType.Int になり、型が食い違っていた。
収集品の納品で直したのと同じ誤りが、交換側にも残っていた。

橙スクリップでの実測により、次が確定した。

  系統を選ぶ  Fire(12, 位置)      0 起点
  種別を選ぶ  Fire(13, タブ)      0 起点
  交換する    Fire(14, index, 数量)
  確認        ShopExchangeItemDialog の Fire(0)
  追加の確認  品によっては SelectYesno の Fire(0)

読み取った index がそのまま発火に使えることも、観測した品揃えと
実際の発火が一致することで確かめられた。

+++ AutoCollector/Automation/ExchangeExecutor.cs
@@ -2050,7 +2050,10 @@ public sealed unsafe class ExchangeExecutor(
             $"（コスト {definition.CurrencyCost} / index {callbackIndex} / アイテム交換画面）");
 
         // 数量は 1 回につき 1 個。まとめ買いは結果の検証が複雑になるため行わない。
-        Callback.Fire(addon, true, InclusionExchangeCommand, callbackIndex, 1);
+        //
+        // 実測は Fire(14, 0u, 1u)。コマンドは Int、index と数量は UInt だった。
+        // int のまま渡すと AtkValueType.Int になり、実測と型が食い違う。
+        Callback.Fire(addon, true, InclusionExchangeCommand, (uint)callbackIndex, 1u);
 
         // 実測では、撃った直後に確認ダイアログが出る。
         // これに答えないと交換は成立しない。
